done welcome screen and waiting screen
