+++ JavaApp/PlayScreen/core/assets/waitingRoom/waitingRoom_.txt
@@ -1,26 +1,54 @@
 waitingRoom_.png
-size: 1922, 1449
+size: 2002, 1427
 format: RGBA8888
 filter: Linear,Linear
 repeat: none
-background
+backgroundWaitingRoom
   rotate: false
   xy: 1, 1
-  size: 1920, 1080
-  orig: 1920, 1080
+  size: 2000, 1125
+  orig: 2000, 1125
   offset: 0, 0
   index: -1
-playerAvailable
-  rotate: true
-  xy: 1, 1083
-  size: 365, 664
-  orig: 365, 664
+connectedInforOff
+  rotate: false
+  xy: 1041, 1235
+  size: 584, 86
+  orig: 584, 86
   offset: 0, 0
   index: -1
-playerNotAvailable
+connectedInforOn
+  rotate: false
+  xy: 1041, 1323
+  size: 583, 86
+  orig: 583, 86
+  offset: 0, 0
+  index: -1
+playBtnOff
+  rotate: false
+  xy: 1041, 1128
+  size: 456, 105
+  orig: 456, 113
+  offset: 0, 8
+  index: -1
+playBtnOn
+  rotate: false
+  xy: 1499, 1128
+  size: 456, 105
+  orig: 456, 114
+  offset: 0, 9
+  index: -1
+playerStatusOff
   rotate: true
-  xy: 667, 1083
-  size: 365, 664
-  orig: 365, 664
+  xy: 1, 1128
+  size: 298, 518
+  orig: 298, 518
+  offset: 0, 0
+  index: -1
+playerStatusOn
+  rotate: true
+  xy: 521, 1128
+  size: 297, 518
+  orig: 297, 518
   offset: 0, 0
   index: -1
