--- JavaApp/PlayScreen/core/assets/welcomeScreen/welcomeScreen_.txt
@@ -1,39 +1,0 @@
-welcomeScreen_-0.png
-size: 1922, 1983
-format: RGBA8888
-filter: Linear,Linear
-repeat: none
-background
-  rotate: false
-  xy: 1, 1
-  size: 1920, 1080
-  orig: 1920, 1080
-  offset: 0, 0
-  index: -1
-IoTLogo
-  rotate: false
-  xy: 1, 1083
-  size: 899, 899
-  orig: 899, 899
-  offset: 0, 0
-  index: -1
-IULogo
-  rotate: false
-  xy: 902, 1083
-  size: 844, 843
-  orig: 886, 885
-  offset: 21, 21
-  index: -1
-
-welcomeScreen_-1.png
-size: 1121, 458
-format: RGBA8888
-filter: Linear,Linear
-repeat: none
-monoLogo
-  rotate: false
-  xy: 1, 1
-  size: 1119, 456
-  orig: 1134, 465
-  offset: 9, 3
-  index: -1
