+++ JavaApp/PlayScreen/core/src/com/monopoly/MonopolyOOP.java
@@ -1,8 +1,9 @@
 package com.monopoly;
 
-
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.monopoly.gamePlay.WaitingRoom;
+import com.monopoly.gamePlay.WelcomeScreen;
 
 public class MonopolyOOP extends Game {
