+++ JavaApp/PlayScreen/desktop/src/com/monopoly/desktop/DesktopLauncher.java
@@ -8,6 +8,7 @@ public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
 
+		// near by 16:9
 		config.foregroundFPS = 60;
 		config.width = 1800;
 		config.height = 1000;
